pyproject.toml
[project]
name = "tradfi-swings"
version = "0.1.0"
description = "Fib confluence S/R tool for TradFi (forex, commodities, indices, stocks) with VIX/DXY market context"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.2",
    "numpy>=1.26",
    "requests>=2.31",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
